requirements.txt
bson
Flask==0.12.2
pymongo==2.8.1
python-dotenv
https://github.com/adebayoj/fairml/archive/master.zip
sklearn 
scikit-learn